refactor(naming): extract apply_naming_to_instance_values to naming_utils
